+++ src/App.css
@@ -22,6 +22,8 @@ a {
     text-decoration: none;
 }
 
+
+
 body {
     background: #f6f5f7;
 	display: flex;
