+++ src/Login/Register.js
@@ -5,9 +5,7 @@ import logo from './logo.png';
 
 class Register extends React.Component {
     
-    constructor(props) {
-        super(props);
-    }
+    
 
 
 render() {
@@ -33,8 +31,6 @@ render() {
             </div>
         </div>
         </div>
-        <div className="footer">
-            <button type="button" className="btn">Register</button></div>
         </div>     
     ); 
     }
