+++ src/Todolist/todolist.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import TodoItem from '../TodoItem/todoitem';
+import 'styled-components';
 
 
 class TodoList extends Component {
